Fix contract
Set cards needed to be at the beginning of round initialization
Tried some things for the contract to automatically initialize a round
by changing the migrations/2_deploy_contracts.js file. Not sure if it
works.
Based on whether the above works or not, contract has to be changed.
Once that is fixed, proper tests can be written